Refactoring of the new pagination
